+++ src/main/java/software/amazon/awssdk/crt/iot/IoTMetricsMetadata.java
@@ -2,10 +2,9 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
  */
-package software.amazon.awssdk.crt.internal;
+package software.amazon.awssdk.crt.iot;
 
 /**
- * @internal
  * A key-value pair for IoT SDK metrics metadata.
  * Metadata entries are appended to the MQTT CONNECT packet username field
  * as part of the Metadata query parameter.
