+++ src/main/java/software/amazon/awssdk/crt/mqtt/MqttClientConnection.java
@@ -17,8 +17,7 @@ import software.amazon.awssdk.crt.mqtt.MqttConnectionConfig;
 import software.amazon.awssdk.crt.mqtt5.Mqtt5Client;
 import software.amazon.awssdk.crt.mqtt5.Mqtt5ClientOptions;
 import software.amazon.awssdk.crt.mqtt5.packets.ConnectPacket;
-import software.amazon.awssdk.crt.internal.IoTDeviceSDKMetrics;
-import software.amazon.awssdk.crt.internal.IoTMetricEncoder;
+import software.amazon.awssdk.crt.iot.IoTDeviceSDKMetrics;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
@@ -166,13 +165,7 @@ public class MqttClientConnection extends CrtResource {
             }
 
             if (!config.getDisableMetrics()) {
-                TlsContext tls = null;
-                if (config.getMqttClient()!=null){
-                    tls = config.getMqttClient().getTlsContext();
-                } else if (config.getMqtt5Client()!=null){
-                    tls = config.getMqtt5Client().getClientOptions().getTlsContext();
-                }
-                IoTDeviceSDKMetrics metrics = IoTMetricEncoder.createMetricsMqtt3(config.getMetrics(), config.getHttpProxyOptions(), tls);
+                IoTDeviceSDKMetrics metrics = IoTDeviceSDKMetrics.createMetricsMqtt3(config);
                 mqttClientConnectionSetMetrics(getNativeHandle(), metrics);
             }
